Moved the router routes into route files
