+++ app/initialize.js
@@ -6,5 +6,6 @@ require('models');
 require('controllers');
 require('views');
 require('router');
+require('routes');
 
 App.initialize();
