+++ app/routes.js
@@ -1,0 +1,4 @@
+// load all your routes here
+
+require('routes/index');
+require('routes/bob');
