+++ app/routes/index.js
@@ -1,0 +1,7 @@
+var App = require('app');
+
+App.IndexRoute = Ember.Route.extend({
+    redirect: function() {
+        this.transitionTo('home');
+    }
+});
